Use React.Strict and stop using v1/auth endpoints
